debug and refactor create recipes code
